Split lnetatmo in sub-modules
